serdes.h: Make Triangle more space efficient.

+++ src/accelerators/cloud.h
@@ -103,7 +103,7 @@ class CloudBVH : public Aggregate {
     using TriangleMeshId = std::pair<uint32_t, uint32_t>;
 
     mutable std::map<uint32_t, Treelet> treelets_;
-    mutable std::map<uint64_t, std::shared_ptr<Primitive>> bvh_instances_;
+    mutable std::map<uint32_t, std::shared_ptr<Primitive>> bvh_instances_;
     mutable std::list<std::unique_ptr<Transform>> transforms_;
     mutable std::map<TriangleMeshId, std::shared_ptr<TriangleMesh>>
         triangle_meshes_;
